\name{packHV-package}
\alias{packHV-package}
\title{packHV}
\alias{packHV}
\docType{package}
\description{
Various useful functions for statisticians: describe data, plot Kaplan-Meier curves with numbers of subjects at risk, compare data sets, display spaghetti-plot, build multi-contingency tables...
}
\details{
\tabular{ll}{
Package: \tab packHV\cr
Type: \tab Package\cr
Version: \tab 1.6\cr
Date: \tab 2013-06-05\cr
License: \tab GPL (>=2) \cr
}
}
\author{
Hugo Varet
}
