\name{packHV-package}
\alias{packHV-package}
\title{packHV}
\alias{packHV}
\docType{package}
\description{
A few usefull functions
}
\details{
\tabular{ll}{
Package: \tab packHV\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2013-02-25\cr
License: \tab GPL (>=2) \cr
}
}
\author{
Hugo Varet
}
