% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packHV-package.r
\docType{package}
\name{packHV-package}
\alias{packHV-package}
\title{A few useful functions for statisticians}
\description{
Various useful functions for statisticians: describe data, plot Kaplan-Meier curves with numbers of subjects at risk, compare data sets, display spaghetti-plot, build multi-contingency tables...
}
\author{
Hugo Varet
}
