% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_multi.table.r
\name{plot_multi.table}
\alias{plot_multi.table}
\title{Plot a multi cross table}
\usage{
plot_multi.table(data, vars, main = "")
}
\arguments{
\item{data}{the data frame in which we can find \code{vars}}

\item{vars}{vector of character string of covariates}

\item{main}{main title of the plot}
}
\value{
None
}
\description{
Plots a multi cross table on a graph
}
\examples{
plot_multi.table(cgd,c("treat","sex","inherit"))
}
\seealso{
\code{\link{multi.table}}
}
\author{
Hugo Varet
}
